+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Extended weekly task selection to allocate partial task hours and report
+  assigned hours per regular and recurring task.
+
 ## [0.3.0] - 2026-07-05
 
 ### Added